reflex/requirements.txt
pywebview>=4.4
pynput>=1.7.6
mss>=9.0
numpy>=1.24
Pillow>=10.0
